backend/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
redis==5.0.1
aio-pika==9.3.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-dotenv==1.0.0
